[#25] refactor: 리뷰반영
