+++ src/test/java/com/ghm/giftcardfleamarket/brand/service/BrandServiceTest.java
@@ -1,6 +1,6 @@
 package com.ghm.giftcardfleamarket.brand.service;
 
-import static com.ghm.giftcardfleamarket.common.utils.constants.Pagination.*;
+import static com.ghm.giftcardfleamarket.common.utils.constants.PageSize.*;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.BDDMockito.*;
